Clean up logic for closing a buffer
Previously, if the user stated that they wanted to save an unsaved
buffer that they were closing, and then cancelled the file selection,
the operation would silently fail, and the file would close without
being saved. Now, the user will be reprompted in that instance until
they either explicitly say "no" or give a valid file name to save as.